cortex/python/requirements.full.txt
lightrag-hku
fastapi
uvicorn
networkx
numpy<2

flask-cors
google-generativeai
python-dotenv
pyautogui
websockets
requests
torch
transformers
accelerate
bitsandbytes
pillow
huggingface_hub
protobuf
sentencepiece
faster-whisper
piper-tts
soundfile
kokoro
moshi
